backend/requirements.txt
fastapi
uvicorn[standard]
yfinance
python-dotenv
sse-starlette
google-generativeai
alpaca-py
requests
groq
openai
